CSV parser and eliminate cycles (#46)
Parser of the input graph of works and contractors is built:

- read the input graph of works and contractors (.csv file);
- preprocess the obtained input graph of works;
- search for loops in the graph;
- if there are loops, the parser breaks them;
- construction of the work graph.

Connections in the work graph can be reconstructed on the basis of historical data. In this case, additional features are possible: there may be links that are represented by experts and that should not be changed when reconstructing information about links, and there may be details of links that can be changed.