Introduction to ScheduledProject (#61)
Now SchedulingPipeline returns ScheduledProject type instead of just Schedule.
It contains Schedule ifself and, in addition, WorkGraph and list[Contractor], that were used to create it.

On this evidence, WorkUnit was removed from ScheduledWork at all. There are some fields from it copying to ScheduledWork on creation.

Now we have an object holds all actual information about schedule and it's references, connections, resources, and so on.